Add sniffer command to Zeta CLI
