+++ setup.py
@@ -2,6 +2,7 @@
 # -*- coding: utf-8 -*-
 
 import setuptools
+
 from zeta import __version__
 
 with open("README.md", "r") as fh:
@@ -24,6 +25,6 @@ setuptools.setup(
         "Operating System :: OS Independent",
     ],
     python_requires='>=3.6',
-    entry_points={"console_scripts": ["zeta = zeta.zeta:run"]},
+    entry_points={"console_scripts": ["zeta = zeta.zeta_cli:run"]},
     include_package_data=True,
     install_requires=["pyyaml"])
